% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\name{expose_sdio}
\alias{expose_sdio}
\title{expose a reader}
\usage{
expose_sdio(platform = "xenium")
}
\description{
expose a reader
}
\examples{
expose_sdio("xenium") # not basilisk-standards compliant -- leaks python reference into R
}
